requirements.txt
asgiref==3.9.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==4.2.23
django-allauth==65.10.0
idna==3.10
mssql-django==1.5
numpy==2.0.2
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
PyMuPDF==1.26.3
pyodbc==5.2.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
